Import original library files
The files have been obtained from the web archive [1]. They were in the
archive ``gbajpeg.zip``. However, this file had a lot of unrelated
files, like GBFS tooling and example files. Only the library files have
been added to this repository.

The library was written by Burton Radons and it is released under the
public domain.

[1] https://web.archive.org/web/20230130121359/http://members.iinet.net.au/~freeaxs/gbacomp/#Displaying%20a%20JPEG%20Image%20on%20the%20GBA